// frontend/src/lib/api/models/dailyPick.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Paris Sportif API
 * API for football predictions and sports betting analysis
 * OpenAPI spec version: 1.0.0
 */
import type { Prediction } from './prediction';

export interface DailyPick {
  rank: number;
  pick_score?: number;
  prediction: Prediction;
}